Revert "REVAMP: Modern Three-Section Layout with Responsive Design - Fix Window Opacity Error"
